python-version/requirements.txt
# Streamlit Cloud Dependencies
streamlit>=1.29.0
streamlit-mic-recorder>=0.0.4
gTTS>=2.4.0
langchain-groq>=0.0.1
python-dotenv>=1.0.0
groq>=0.9.0
requests>=2.31.0
fpdf>=1.7.2
